requirements.txt
astroid==2.4.0
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
coverage==5.0a1
cryptography==2.9
Django==2.2.9
django-azure-ad-auth==1.3.1
django-cors-headers==3.2.1
django-jenkins==0.19.0
djangorestframework==3.11.0
Flask==1.0.2
gunicorn==19.9.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jslint==0.6.0
lazy-object-proxy==1.4.3
lxml==4.5.0
mailjet-rest==1.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.5
packaging==20.3
pandas==1.0.4
pep8==1.7.1
Pillow==7.1.2
pluggy==0.13.1
psycopg2-binary
py==1.8.1
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.0
pylint-runner==0.5.4
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.44
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
unittest-xml-reporting==3.0.2
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==0.16.0
wrapt==1.11.2
xmlrunner==1.7.7
zipp==3.1.0

#psycopg2~=2.8.5
PyYAML~=5.1.1